+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 - Added test for negative response and introduce error=None if response status code is 200 (success)
 - Added simple query interface
 - Run queries and display results on the web interface
+- Add 3 level of authentication when creating datasets and fix tests accordingly